+++ src/web/logs.rs
@@ -1,5 +1,5 @@
-use std::collections::HashMap;
 
+use std::collections::HashMap;
 use chrono::NaiveDate;
 
 use super::about;
@@ -328,54 +328,7 @@ impl Logs {
     }
 }
 
-fn derive_entries(storage_state: &StorageState) -> HashMap<NaiveDate, Vec<Entry>> {
-    let mut entries: HashMap<NaiveDate, Vec<Entry>> = HashMap::new();
-    for m in &storage_state.mood_readings {
-        let d = entry_date(&Entry::Mood(m.clone()));
-        if let Some(e) = entries.get_mut(&d) {
-            e.push(Entry::Mood(m.clone()))
-        } else {
-            entries.insert(d, vec![Entry::Mood(m.clone())]);
-        }
-    }
-    for s in &storage_state.sleep_entries {
-        let d = entry_date(&Entry::Sleep(s.clone()));
-        if let Some(e) = entries.get_mut(&d) {
-            e.push(Entry::Sleep(s.clone()))
-        } else {
-            entries.insert(d, vec![Entry::Sleep(s.clone())]);
-        }
-    }
-    for m in &storage_state.meds {
-        let d = entry_date(&Entry::Meds(m.clone()));
-        if let Some(e) = entries.get_mut(&d) {
-            e.push(Entry::Meds(m.clone()))
-        } else {
-            entries.insert(d, vec![Entry::Meds(m.clone())]);
-        }
-    }
-    for n in &storage_state.notes {
-        let d = entry_date(&Entry::Note(n.clone()));
-        if let Some(e) = entries.get_mut(&d) {
-            e.push(Entry::Note(n.clone()))
-        } else {
-            entries.insert(d, vec![Entry::Note(n.clone())]);
-        }
-    }
 
-    entries
-}
-
-fn entry_date(e: &Entry) -> NaiveDate {
-    let date = js_sys::Date::new(&JsValue::from_f64(e.timestamp() as f64));
-
-    NaiveDate::from_ymd_opt(
-        date.get_full_year() as i32,
-        date.get_month() as u32 + 1,
-        date.get_date() as u32,
-    )
-    .unwrap()
-}
 
 fn format_timestamp(epoch_millis_utc: u64) -> String {
     let date = js_sys::Date::new(&JsValue::from_f64(epoch_millis_utc as f64));
